+++ src/commands/openWorkingFile.ts
@@ -50,7 +50,7 @@ export class OpenWorkingFileCommand extends ActiveEditorCommand {
             const e = await openEditor(args.uri, { ...args.showOptions, rethrow: true });
             if (args.annotationType === undefined) return e;
 
-            return Container.annotations.showAnnotations(e!, args.annotationType, args.line);
+            return Container.fileAnnotations.show(e!, args.annotationType, args.line);
         }
         catch (ex) {
             Logger.error(ex, 'OpenWorkingFileCommand');